backend/backend/requirements/dev.txt
-r base.txt

# Lint / format / type
ruff==0.5.1
black==24.4.2
mypy==1.10.1
django-stubs[compatible-mypy]==5.0.2
djangorestframework-stubs==3.15.0

# Testing
pytest==8.2.2
pytest-django==4.8.0
pytest-cov==5.0.0
pytest-asyncio==0.23.7
factory-boy==3.3.0
faker==26.0.0

# Dev tooling
django-debug-toolbar==4.4.6
django-extensions==3.2.3
ipython==8.26.0
